requirements-prod.txt
# For production deployment (IPv4 forced for Render compatibility)
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-CORS==4.0.0
psycopg2-binary>=2.9.5
requests==2.31.0
python-dotenv==1.0.0
gunicorn==21.2.0
dnspython==2.4.2